+++ src/pynescript/ast/evaluator/expressions.py
@@ -1,0 +1,102 @@
+from __future__ import annotations
+
+import itertools
+import operator
+
+from typing import Any
+
+from pynescript.ast import node as ast
+
+
+class ExpressionEvaluator:
+
+    def visit_BoolOp(self, node: ast.BoolOp):
+        if isinstance(node.op, ast.And):
+            return all(self.visit(value) for value in node.values)
+        if isinstance(node.op, ast.Or):
+            return any(self.visit(value) for value in node.values)
+        msg = f"unexpected node operator: {node.op}"
+        raise ValueError(msg)
+
+    def visit_BinOp(self, node: ast.BinOp):
+        left = self.visit(node.left)
+        right = self.visit(node.right)
+        if isinstance(node.op, ast.Add):
+            return operator.add(left, right)
+        elif isinstance(node.op, ast.Sub):
+            return operator.sub(left, right)
+        elif isinstance(node.op, ast.Mult):
+            return operator.mul(left, right)
+        elif isinstance(node.op, ast.Div):
+            return operator.truediv(left, right)
+        elif isinstance(node.op, ast.Mod):
+            return operator.mod(left, right)
+        else:
+            msg = f"Unsupported binary operator: {type(node.op)}"
+            raise NotImplementedError(msg)
+
+    def visit_UnaryOp(self, node: ast.UnaryOp):
+        if isinstance(node.op, ast.Not):
+            return operator.not_(self.visit(node.operand))
+        if isinstance(node.op, ast.UAdd):
+            return operator.pos(self.visit(node.operand))
+        if isinstance(node.op, ast.USub):
+            return operator.neg(self.visit(node.operand))
+        msg = f"unexpected node operator: {node.op}"
+        raise ValueError(msg)
+
+    def visit_Conditional(self, node: ast.Conditional) -> Any:
+        test_result = self.visit(node.test)
+        if test_result:
+            return self.visit(node.body)
+        else:
+            return self.visit(node.orelse)
+
+    def visit_Compare(self, node: ast.Compare) -> Any:
+        comparator_list = [node.left, *node.comparators]
+        comparators = map(self.visit, comparator_list)
+        compare_ops = [self.visit(op) for op in node.ops]
+        comparator_pairs = list(itertools.pairwise(comparators))
+        pairs = zip(compare_ops, comparator_pairs, strict=True)
+        for op, (left, right) in pairs:
+            if not op(left, right):
+                return False
+        return True
+
+    def visit_Eq(self, _node: ast.Eq):
+        return operator.eq
+
+    def visit_NotEq(self, _node: ast.NotEq):
+        return operator.ne
+
+    def visit_Lt(self, _node: ast.Lt):
+        return operator.lt
+
+    def visit_LtE(self, _node: ast.LtE):
+        return operator.le
+
+    def visit_Gt(self, _node: ast.Gt):
+        return operator.gt
+
+    def visit_GtE(self, _node: ast.GtE):
+        return operator.ge
+
+    def visit_Call(self, node: ast.Call):
+        func = self.visit(node.func)
+        args = []
+
+        kwargs = {}
+
+        for arg in node.args:
+            if arg.name:
+                kwargs[arg.name] = self.visit(arg.value)
+
+            else:
+                args.append(self.visit(arg.value))
+
+        # Handle built-in functions
+        if isinstance(func, str):
+            return self._call_builtin(func, args)
+        else:
+            # For now, assume func is callable
+            return func(*args, **kwargs)
